removed engine.h from polyring.hpp
