+++ M2/Macaulay2/e/LLL.cpp
@@ -1,11 +1,12 @@
 // Copyright 1998  Michael E. Stillman
 
 #include "LLL.hpp"
+
 #include "relem.hpp"
 #include "matrix.hpp"
-#include "frac.hpp"
 #include "text-io.hpp"
 #include "interrupted.hpp"
+
 bool LLLoperations::checkThreshold(ring_elem num, ring_elem den)
 {
   // Makes sure that 1/4 < num/den <= 1
