+++ M2/Macaulay2/e/engine.h
@@ -11,6 +11,7 @@
 #include "rand.h" // engine interface for random numbers
 
 #include "interface/monomial-ordering.h"
+#include "interface/computation.h"
 
 #if defined(__cplusplus)
 class Monomial;
@@ -1668,28 +1669,6 @@ extern "C" {
   /**************************************************/
   /**** Groebner basis and resolution routines ******/
   /**************************************************/
-
-  enum ComputationStatusCode {
-    /* Keep this enum in sync with RawStatusCodes in Macaulay2/m2/gb.m2 */
-    COMP_NEED_RESIZE = 1,         /* need resize */
-    COMP_ERROR = 2,               /* error */
-    COMP_INTERRUPTED = 3,         /* interrupted */
-    COMP_NOT_STARTED = 4,         /* not started */
-    COMP_INITIAL_STOP = 5,        /* StopBeforeComputation */
-    COMP_DONE = 6,                /* done */
-    COMP_DONE_DEGREE_LIMIT = 7,   /* DegreeLimit */
-    COMP_DONE_LENGTH_LIMIT = 8,   /* LengthLimit */
-    COMP_DONE_SYZYGY_LIMIT = 9,   /* SyzygyLimit */
-    COMP_DONE_PAIR_LIMIT = 10,    /* PairLimit */
-    COMP_DONE_GB_LIMIT = 11,      /* BasisElementLimit */
-    COMP_DONE_SYZ_LIMIT = 12,     /* SyzygyLimit */
-    COMP_DONE_CODIM = 13,         /* CodimensionLimit */
-    COMP_DONE_MIN_GENS = 14,      /* StopWithMinimalGenerators */
-    COMP_DONE_STEPS = 15,         /* StepLimit */
-    COMP_DONE_SUBRING_LIMIT = 16, /* SubringLimit */
-    COMP_COMPUTING = 17,          /* computing */
-    COMP_OVERFLOWED = 18,         /* overflowed */
-  };
 
   enum StrategyValues {
     STRATEGY_LONGPOLYNOMIALS = 1,
