+++ M2/Macaulay2/e/interface/polyroots.hpp
@@ -1,9 +1,9 @@
-
 #ifndef _polyroots_hpp_
 #define _polyroots_hpp_
 
-#include "relem.hpp"
 #include "polyring.hpp"
+
+#include "relem.hpp"
 #include "aring-CC.hpp"
 #include "aring-CCC.hpp"
 #include "aring-glue.hpp"
@@ -18,11 +18,13 @@ typedef M2::ConcreteRing<M2::ARingCCC> RingCCC;
 typedef cc_struct mpfc_t[1];
 typedef cc_ptr mpfc_ptr;
 
-engine_RawRingElementArrayOrNull rawRoots(const RingElement *p,
-                                          long prec,
-                                          int unique);
+/* FIXME: this is also defined in engine.h
+extern "C" engine_RawRingElementArrayOrNull rawRoots(const RingElement *p,
+                                                     long prec,
+                                                     int unique);
+*/
 
-#endif
+#endif /* _polyroots_hpp_ */
 
 // Local Variables:
 // compile-command: "make -C $M2BUILDDIR/Macaulay2/e  "
