+++ M2/Macaulay2/e/matrix.hpp
@@ -8,6 +8,7 @@
 #include "monideal.hpp"
 
 #include <vector>
+
 class MatrixConstructor;
 
 /**
